# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Portal Manajemen Proyek API
  version: 1.0.0
  description: REST API untuk autentikasi, RBAC, CRUD proyek, tugas, dashboard, risiko, time tracking, file, dan komentar.
servers:
  - url: http://localhost:4000/api
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: Daftar tugas sesuai role, status, projectId, dan rentang tanggal
      parameters:
        - in: query
          name: projectId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [todo, in_progress, done]
        - in: query
          name: fromDate
          schema:
            type: string
            format: date
        - in: query
          name: toDate
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daftar tugas terfilter
    post:
      tags: [Tasks]
      summary: Membuat tugas, hanya Project Manager
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Tugas dibuat
  /tasks/{id}:
    put:
      tags: [Tasks]
      summary: Project Manager mengubah tugas, Developer mengubah status/progress tugasnya
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "200":
          description: Tugas diperbarui
    delete:
      tags: [Tasks]
      summary: Menghapus tugas, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Tugas dihapus
...
  /task-dependencies/{id}:
    put:
      tags: [Dashboard]
      summary: Mengubah dependensi tugas, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Dependensi diperbarui
    delete:
      tags: [Dashboard]
      summary: Menghapus dependensi tugas, hanya Project Manager
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Dependensi dihapus
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      in: path
      name: id
      required: true
      schema:
        type: integer
  schemas:
...
    TaskInput:
      type: object
      required: [project_id, name]
      properties:
        project_id:
          type: integer
        name:
          type: string
        description:
          type: string
        assigned_to:
          type: integer
        status:
          type: string
          enum: [todo, in_progress, done]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        due_date:
          type: string
          format: date